Bot discord/requirements.txt
discord.py
asyncio
youtube-dl
google-api-python-client